+++ poly_fill.hh
@@ -6,7 +6,7 @@
 #include "geom.hh"
 #include "image.hh"
 
-void fillPoly(const Poly& poly, Image<uint8_t>& outimg, uint8_t val);
+void fillPoly(const Poly& poly, Image<float>& outimg, float val);
 
 void fillPoly2(const PolyVec& detpoly, const PolyVec& maskpoly,
                Image<float>& outimg, float fillval);
